minor reorganization, changed readme
