+++ Controller/src/virtual_robot/controller/Game.java
@@ -1,4 +1,4 @@
-package virtual_robot.config;
+package virtual_robot.controller;
 
 import java.lang.annotation.Annotation;
 import java.util.ArrayList;
@@ -9,11 +9,10 @@ import java.util.Set;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Group;
 import javafx.scene.layout.Pane;
-import javafx.scene.layout.StackPane;
 import org.dyn4j.dynamics.Body;
 import org.dyn4j.world.World;
 import org.reflections.Reflections;
-import virtual_robot.controller.*;
+import virtual_robot.config.Config;
 
 public abstract class Game {
