+++ Controller/src/virtual_robot/game_elements/classes/WobbleGoal.java
@@ -5,9 +5,7 @@ import javafx.scene.Group;
 import javafx.scene.shape.Circle;
 import org.dyn4j.collision.CategoryFilter;
 import org.dyn4j.dynamics.Body;
-import org.dyn4j.dynamics.BodyFixture;
-import org.dyn4j.geometry.MassType;
-import virtual_robot.config.UltimateGoal;
+import virtual_robot.games.UltimateGoal;
 import virtual_robot.controller.*;
 import virtual_robot.dyn4j.Dyn4jUtil;
 import virtual_robot.dyn4j.FixtureData;
