+++ Controller/src/virtual_robot/games/UltimateGoal.java
@@ -1,4 +1,4 @@
-package virtual_robot.config;
+package virtual_robot.games;
 
 import java.util.*;
 
@@ -7,6 +7,7 @@ import org.dyn4j.dynamics.BodyFixture;
 import org.dyn4j.geometry.Vector2;
 import org.dyn4j.world.NarrowphaseCollisionData;
 import org.dyn4j.world.listener.CollisionListenerAdapter;
+import virtual_robot.controller.Game;
 import virtual_robot.controller.VirtualGameElement;
 import virtual_robot.game_elements.classes.Ring;
 import virtual_robot.game_elements.classes.WobbleGoal;
