+++ Controller/src/virtual_robot/robots/ControlsElements.java
@@ -1,6 +1,6 @@
 package virtual_robot.robots;
 
-import virtual_robot.config.Game;
+import virtual_robot.controller.Game;
 
 public interface ControlsElements {
     void preloadElements(Game game);
