+++ Controller/src/virtual_robot/robots/classes/BetaBot.java
@@ -14,8 +14,8 @@ import org.dyn4j.geometry.Rectangle;
 import org.dyn4j.geometry.Vector2;
 import org.dyn4j.world.NarrowphaseCollisionData;
 import org.dyn4j.world.listener.CollisionListenerAdapter;
-import virtual_robot.config.Game;
-import virtual_robot.config.UltimateGoal;
+import virtual_robot.controller.Game;
+import virtual_robot.games.UltimateGoal;
 import virtual_robot.controller.BotConfig;
 import virtual_robot.controller.Filters;
 import virtual_robot.controller.VirtualField;
